Fix some warnings
Updating pods gets rid of some "-pie" warnings.
Some activitiy indicators were using a style only available in iOS 13, and those were replaced with CircleProgressView.
Submission comments used a fx style only in iOS 13, that was reverted to an older style.
Some CoreData relationships needed an inverse, even if we never expose it with a property.

refs: none
affects: none
release note: none
