+++ Core/Core/Models/Database.xcdatamodeld/Database.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="15508" systemVersion="19D76" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="16119" systemVersion="19E287" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="Activity" representedClassName=".Activity" syncable="YES">
         <attribute name="canvasContextIDRaw" optional="YES" attributeType="String"/>
         <attribute name="createdAt" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
@@ -139,8 +139,9 @@
         <attribute name="participantIDsRaw" optional="YES" attributeType="String"/>
         <relationship name="attachmentsRaw" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="File" inverseName="conversationMessage" inverseEntity="File"/>
         <relationship name="conversation" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="Conversation" inverseName="messagesRaw" inverseEntity="Conversation"/>
-        <relationship name="forwardedRaw" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="ConversationMessage"/>
+        <relationship name="forwardedRaw" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="ConversationMessage" inverseName="parentMessage" inverseEntity="ConversationMessage"/>
         <relationship name="mediaComment" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="MediaComment" inverseName="conversationMessage" inverseEntity="MediaComment"/>
+        <relationship name="parentMessage" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="ConversationMessage" inverseName="forwardedRaw" inverseEntity="ConversationMessage"/>
     </entity>
     <entity name="ConversationParticipant" representedClassName="Core.ConversationParticipant" syncable="YES">
         <attribute name="avatarURL" optional="YES" attributeType="URI"/>
@@ -221,7 +222,7 @@
         <attribute name="userID" optional="YES" attributeType="String"/>
         <relationship name="course" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="Course" inverseName="enrollments" inverseEntity="Course"/>
         <relationship name="grades" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="Grade" inverseName="enrollment" inverseEntity="Grade"/>
-        <relationship name="observedUser" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="User"/>
+        <relationship name="observedUser" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="User" inverseName="observerEnrollments" inverseEntity="User"/>
         <relationship name="user" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="User" inverseName="enrollments" inverseEntity="User"/>
     </entity>
     <entity name="ExternalTool" representedClassName="Core.ExternalTool" syncable="YES">
@@ -617,6 +618,7 @@
         <attribute name="shortName" attributeType="String"/>
         <attribute name="sortableName" attributeType="String"/>
         <relationship name="enrollments" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="Enrollment" inverseName="user" inverseEntity="Enrollment"/>
+        <relationship name="observerEnrollments" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="Enrollment" inverseName="observedUser" inverseEntity="Enrollment"/>
     </entity>
     <entity name="UserProfile" representedClassName="Core.UserProfile" syncable="YES">
         <attribute name="avatarURL" optional="YES" attributeType="URI"/>
@@ -644,7 +646,7 @@
         <element name="ConferenceRecording" positionX="-531" positionY="-9" width="128" height="148"/>
         <element name="ContextColor" positionX="-177.3046875" positionY="-36" width="128" height="73"/>
         <element name="Conversation" positionX="-540" positionY="-18" width="128" height="238"/>
-        <element name="ConversationMessage" positionX="-522" positionY="0" width="128" height="178"/>
+        <element name="ConversationMessage" positionX="-522" positionY="0" width="128" height="208"/>
         <element name="ConversationParticipant" positionX="-531" positionY="-9" width="128" height="118"/>
         <element name="Course" positionX="-365.20703125" positionY="2.984375" width="128" height="283"/>
         <element name="CourseSection" positionX="-540" positionY="-18" width="128" height="133"/>
@@ -680,7 +682,7 @@
         <element name="Tab" positionX="108.015625" positionY="177.67578125" width="128" height="165"/>
         <element name="Todo" positionX="-540" positionY="-18" width="128" height="148"/>
         <element name="TTL" positionX="-173.30859375" positionY="-128.48046875" width="128" height="75"/>
-        <element name="User" positionX="-540" positionY="-27" width="128" height="208"/>
+        <element name="User" positionX="-540" positionY="-27" width="128" height="223"/>
         <element name="UserProfile" positionX="-540" positionY="-18" width="128" height="148"/>
         <element name="UserSettings" positionX="-540" positionY="-27" width="128" height="90"/>
     </elements>
