+++ Parent/Parent/IBDesignables/IBDesignables.swift
@@ -21,9 +21,11 @@ import Core
 @IBDesignable
 class AvatarGroupView: Core.AvatarGroupView {}
 class AvatarView: Core.AvatarView {}
+class CoreWebView: Core.CoreWebView {}
 class DividerView: Core.DividerView {}
 class DynamicButton: Core.DynamicButton {}
 class DynamicLabel: Core.DynamicLabel {}
 class DynamicTextField: Core.DynamicTextField {}
+class EmptyView: Core.EmptyView {}
 class FloatingButton: Core.FloatingButton {}
 class IconView: Core.IconView {}
