+++ Student/Student/IBDesignables/IBDesignables.swift
@@ -1,6 +1,6 @@
 //
 // This file is part of Canvas.
-// Copyright (C) 2019-present  Instructure, Inc.
+// Copyright (C) 2020-present  Instructure, Inc.
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
@@ -16,8 +16,16 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
-import Foundation
 import Core
 
 @IBDesignable
+class AvatarGroupView: Core.AvatarGroupView {}
+class AvatarView: Core.AvatarView {}
+class CoreWebView: Core.CoreWebView {}
 class DividerView: Core.DividerView {}
+class DynamicButton: Core.DynamicButton {}
+class DynamicLabel: Core.DynamicLabel {}
+class DynamicTextField: Core.DynamicTextField {}
+class EmptyView: Core.EmptyView {}
+class FloatingButton: Core.FloatingButton {}
+class IconView: Core.IconView {}
