+++ Student/Student/Submissions/SubmissionComments/SubmissionCommentTextCell.swift
@@ -30,7 +30,6 @@ class SubmissionCommentTextCell: UITableViewCell {
     override func awakeFromNib() {
         super.awakeFromNib()
         backgroundColor = .named(.backgroundLightest)
-        chatBubbleView?.image = UIImage(named: "chatBubble", in: Bundle.core, compatibleWith: nil)
     }
 
     func update(comment: SubmissionComment) {
