+++ rn/Teacher/ios/Teacher/IBDesignables/IBDesignables.swift
@@ -1,6 +1,6 @@
 //
 // This file is part of Canvas.
-// Copyright (C) 2018-present  Instructure, Inc.
+// Copyright (C) 2019-present  Instructure, Inc.
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
@@ -19,4 +19,13 @@
 import Core
 
 @IBDesignable
+class AvatarGroupView: Core.AvatarGroupView {}
+class AvatarView: Core.AvatarView {}
+class CoreWebView: Core.CoreWebView {}
+class DividerView: Core.DividerView {}
+class DynamicButton: Core.DynamicButton {}
+class DynamicLabel: Core.DynamicLabel {}
+class DynamicTextField: Core.DynamicTextField {}
+class EmptyView: Core.EmptyView {}
+class FloatingButton: Core.FloatingButton {}
 class IconView: Core.IconView {}
